Crate secgate

Macros

secgate_prelude

Required to put in your source if you call any secure gates.

Structs

Arguments

Arguments that will be passed to the secure call. Concrete versions of this are generated by the macro.

GateCallInfo

Return

Return value to be filled by the secure call. Concrete versions of this are generated by the macro.

SecGateInfo

A struct of information about a secure gate. These are auto-generated by the crate::secure_gate macro, and stored in a special ELF section (.twz_secgate_info) as an array. The dynamic linker and monitor can then use this to easily enumerate gates.

Enums

SecGateReturn

Enum of possible return codes, similar to Result, but with specific variants of possible failures of initializing or invoking the secure gate call.

Constants

SECGATE_TRAMPOLINE_ALIGN

Minimum alignment of secure trampolines.

Traits

Crossing

An auto trait that limits the types that can be send across to another compartment. These are:

Type Aliases

RawSecGateInfo

Non-generic and non-pointer-based SecGateInfo, for use during dynamic linking.

Attribute Macros

secure_gate